refactor: extract quota scheduler policy
