+++ loopx/policies/__init__.py
@@ -1,0 +1,6 @@
+"""Pure control-plane policy helpers.
+
+The modules in this package should stay side-effect free and callable from hot
+path builders such as quota/status without depending back on those builders.
+"""
+
